more debugging/cleaning

+++ pipeline.py
@@ -8,12 +8,13 @@ import os
 import sys
 import time
 import random
+import subprocess
 
 # Simulate a full workflow.
 # Folder where bamfile, bioanalyzer profile, genome and gtf are in
-folder_in = 'test_data'
-gtf = os.path.join(folder_in, 'Homo_sapiens.GRCh38.83_chr9.gtf')
-genome = os.path.join(folder_in, 'Homo_sapiens.GRCh38.dna.chromosome.9.fa')
+folder_in = './'
+gtf = os.path.join(folder_in, 'Homo_sapiens.GRCh38.83.gtf')
+genome = os.path.join(folder_in, 'GRCh38.primary_assembly.genome.fa')
 
 # Create output directory for storing everything
 folder_out = os.path.join(folder_in, 'output')
@@ -24,12 +25,28 @@ except Exception:
 
 ### 1. Extract utr information from gtf file
 print ("extracting 3'UTR information ...", end=" ", flush=True)
-start_time = time.time()
-old_stdout = sys.stdout
-sys.stdout = open(os.path.join(folder_out, 'utr_annotation.bed'), 'w')
-extract_three_prime_utr_information(gtf, bed_name_attributes = ["gene_name"])
-sys.stdout = old_stdout
-print ('done [', round(time.time() - start_time, 2), 'seconds ]')
+if os.path.isfile(os.path.join(folder_out, 'utr_annotation.bed')):
+    print ('skipping [ file already exists ]')
+else:
+    start_time = time.time()
+    old_stdout = sys.stdout
+    sys.stdout = open(os.path.join(folder_out, 'utr_annotation_temp.bed'), 'w')
+    extract_three_prime_utr_information(gtf, bed_name_attributes = ["gene_name"])
+    sys.stdout = old_stdout
+    print ('done [', round(time.time() - start_time, 2), 'seconds ]')
+
+    ### 1.1 Clean utr from haplotypes and junk chromosomes
+    with open(os.path.join(folder_out, 'utr_annotation_temp.bed'), 'r') as fin, open(os.path.join(folder_out, 'utr_annotation_unsorted.bed'), 'w') as fout:
+        for line in fin:
+            if line.startswith('chrGL') or line.startswith('chrKI'):
+                continue
+            else:
+                fout.write(line)
+    os.remove(os.path.join(folder_out, 'utr_annotation_temp.bed'))
+
+    ### 1.2 Sort the utr file alphabetically
+    subprocess.call('sort -V output/utr_annotation_unsorted.bed > output/utr_annotation.bed', shell=True)
+    os.remove(os.path.join(folder_out, 'utr_annotation_unsorted.bed'))
 
 ### 2. Extract polyA intervals from genome
 print ('extracting polyA intervals from genome ...', end=" ", flush=True)
@@ -44,11 +61,14 @@ else:
 
 ### 3. Add gene information to polyA intervals
 print ('adding gene annotation to pAi intervals ...', end=" ", flush=True)
-start_time = time.time()
-annotate_pAi_with_gene(os.path.join(folder_out, 'pAi.bed'), 
-                       os.path.join(folder_out, 'utr_annotation.bed'))
-os.rename('pAi_gene.bed', os.path.join(folder_out, 'pAi_gene.bed'))
-print ('done [', round(time.time() - start_time, 2), 'seconds ]')
+if os.path.isfile(os.path.join(folder_out, 'pAi_gene.bed')):
+    print ('skipping [ file already exists ]')
+else:
+    start_time = time.time()
+    annotate_pAi_with_gene(os.path.join(folder_out, 'pAi.bed'), 
+                           os.path.join(folder_out, 'utr_annotation.bed'))
+    os.rename('pAi_gene.bed', os.path.join(folder_out, 'pAi_gene.bed'))
+    print ('done [', round(time.time() - start_time, 2), 'seconds ]')
 
 ### 4. Merge polyA intervals with 3'UTRs into a dictionary
 print ("merging polyA intervals with 3'UTR ...", end=" ", flush=True)
@@ -73,37 +93,73 @@ print ('done [', round(time.time() - start_time, 2), 'seconds ]')
 print ('reading bamfile into memory ...', end=" ", flush=True)
 start_time = time.time()
 bamfile = defaultdict(list)
-with gzip.open(os.path.join(folder_in, 'ds_012_50fix_bamfile.txt.gz'), 'rt') as f:
+with gzip.open(os.path.join(folder_in, 'ds_012_50fix_resequenced_three_mismatches.txt.gz'), 'rt') as f:
     for columns in (row.strip().split() for row in f):
         gene = columns[12][8:]
         bamfile[gene].append([columns[3], columns[11], columns[18]])
 print ('done [', round(time.time() - start_time, 2), 'seconds ]')
 
 ### 7. Collapsing PCR duplicates
-print ('collapsing PCR duplicates ...', end=" ", flush=True)
-start_time = time.time()
+#print ('collapsing PCR duplicates ...', end=" ", flush=True)
+#start_time = time.time()
 #for gene in bamfile:
 #    temp_list = bamfile[gene]
 #    temp_list.sort()
 #    bamfile[gene] = list(temp_list for temp_list,_ in itertools.groupby(temp_list))
-print ('done [', round(time.time() - start_time, 2), 'seconds ]')
+#print ('done [', round(time.time() - start_time, 2), 'seconds ]')
 
 ### 8. Estimate tail lengths per gene.
 # focus on particular genes as examples (single 3'UTRs)
 print ('setting up a tail range of', end=" ")
-tail_range = tail_length_range(10, 500, 60)
+tail_range = tail_length_range(10, 550, 30)
 for length in tail_range:
     print (length, end=" ")
 print ('\n')   
 
+### 9. Read all single UTR genes with no pAi in the UTRs
+genes = []
+with open('single_utr_no_pAi_genes.txt', 'r') as f:
+    for line in f:
+        genes.append(line.rstrip())
+
+### 10. iterate over all genes and predict tails
+with open ('tail_lengths.txt', 'w') as results:
+    for index in range(len(genes)):
+        gene = genes[index]
+        print ('estimating polyA tail length for gene', gene, '...', end=" ")
+        reads = []
+        for item in bamfile[gene]:
+            if (int(pAi_full[gene][0]['start']) - int(item[0]) <= max(f_size)):
+                reads.append(int(item[0]))
+        #reads = [ reads[i] for i in sorted(random.sample(range(len(reads)), 100)) ]
+        # Put threshold for number of reads required
+        if len(reads) < 100:
+            print ('not enough reads for analysis [', len(reads), ']')
+            continue
+        print (len(reads), 'reads will be used for the analysis ...', end=" ")
+        start_time = time.time()
+        probs = estimate_poly_tail_length(reads, tail_range, pAi_full[gene],
+                                          0, f_size, f_prob, False)
+        print ('done [', round(time.time() - start_time, 2), 'seconds ]')
+        results.write(gene + ',' + str(probs) + '\n')
+
+
+
+
+
+
+
+
+sys.exit()
+
 # select the gene here
-gene = 'ANP32B'
+gene = 'CKS2'
 reads = []
 for item in bamfile[gene]:
     if (int(pAi_full[gene][0]['start']) - int(item[0]) <= max(f_size)):
         reads.append(int(item[0]))
 
-# reads = [ reads[i] for i in sorted(random.sample(range(len(reads)), 200)) ]
+reads = [ reads[i] for i in sorted(random.sample(range(len(reads)), 1000)) ]
 
 print (len(reads), 'reads will be used for the analysis')
